refactor: make distribute navbar into a separate component
