+++ web/src/components/distribute/navbar/nav-link.tsx
@@ -1,0 +1,25 @@
+import { Link } from "@tanstack/react-router";
+
+import { cn } from "@/lib/utils";
+
+export function NavLink({
+  href,
+  children,
+  active,
+}: {
+  href: string;
+  children: React.ReactNode;
+  active: boolean;
+}) {
+  return (
+    <Link
+      href={href}
+      className={cn(
+        "flex items-center rounded-lg px-3 py-2 text-sm font-medium transition-colors hover:bg-accent hover:text-accent-foreground",
+        active ? "bg-accent text-accent-foreground" : "text-muted-foreground",
+      )}
+    >
+      {children}
+    </Link>
+  );
+}
